--- .cargo/config.toml
@@ -1,9 +1,0 @@
-[build]
-rustflags = []
-
-# see https://pyo3.rs/main/building_and_distribution.html#macos
-[target.x86_64-apple-darwin]
-rustflags = ["-C", "link-arg=-undefined", "-C", "link-arg=dynamic_lookup"]
-
-[target.aarch64-apple-darwin]
-rustflags = ["-C", "link-arg=-undefined", "-C", "link-arg=dynamic_lookup"]
