+++ crates/karva/Cargo.toml
@@ -1,6 +1,7 @@
 [package]
 name = "karva"
 version = "0.0.12"
+
 edition = { workspace = true }
 rust-version = { workspace = true }
 homepage = { workspace = true }
@@ -10,16 +11,10 @@ authors = { workspace = true }
 license = { workspace = true }
 
 [dependencies]
-karva_core = { workspace = true }
+karva_cli = { workspace = true }
 
-clap = { workspace = true, features = ["wrap_help"] }
-anyhow = { workspace = true }
-argfile = { workspace = true }
-camino = { workspace = true }
-chrono = { workspace = true }
-colored = { workspace = true }
-tracing = { workspace = true, features = ["release_max_level_debug"] }
-tracing-subscriber = { workspace = true, features = ["env-filter", "fmt"] }
-tracing-flame = { workspace = true }
-tracing-tree = { workspace = true }
-wild = { workspace = true }
+pyo3 = { workspace = true }
+
+[features]
+extension-module = ["pyo3/extension-module"]
+default = ["extension-module"]
