+++ crates/karva_core/Cargo.toml
@@ -1,7 +1,6 @@
 [package]
 name = "karva_core"
-version = "0.0.1"
-publish = false
+version = "0.0.0"
 
 edition = { workspace = true }
 rust-version = { workspace = true }
