--- crates/karva_python/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "karva_python"
-version = "0.0.1"
-publish = false
-
-edition = { workspace = true }
-rust-version = { workspace = true }
-homepage = { workspace = true }
-documentation = { workspace = true }
-repository = { workspace = true }
-authors = { workspace = true }
-license = { workspace = true }
-
-[dependencies]
-karva = { workspace = true }
-
-pyo3 = { workspace = true }
-
-[features]
-extension-module = ["pyo3/extension-module"]
-default = ["extension-module"]
